// src/app/shared/models/reservoir-info.ts
export class ReservoirInfo {
  administrator?: number;
  checkfloodlevel?: number;
  deadwaterlevel?: number;
  designfloodlevel?: number;
  floodlimitlevel?: number;
  id?: number;
  inspectionperson?: number;
  latitude?: number;
  longitude?: number;
  managementunit?: number;
  name?: string;
  normalstoragelevel?: number;
  technicalperson?: number;
}